feat: ingredient list redesign — summary cards + portaled bottom sheet (#144)
Replace dense inline IngredientRowEditor with static summary cards (name
+ kcal + P/C/F chips) in both EntryEditor and MealBuilder. Tapping a card
or the "+ Add ingredient" button opens a reusable IngredientSheet portaled
to document.body via createPortal, ensuring it layers above inline chat
proposals, MealBuilder's own sheet, and any other stacking context.
